Don't change anchor links to use the history API if the router doesn't have a match (#4)
The use case for this is integrating `router-slot` into a server-side routed web app where we're proxy-passing from the same parent URL. ie: `foo.com/about` and `foo.com/products` where `/about` is using `router-slot`, but `/products` is not. Anchor links to `/products` will currently try to be handled by `router-slot` because it's a relative link, even if you don't have a route for it.

With this change, if you don't have a catch-all route in `router-slot`, it will allow the browser to redirect the link to the server-side routed `/products` URL.

If the app has a catch-all route, it will still be handled client-side.